Changes to password policies section.
Removed all of the old content about specifying password policies in a
machine-readable format. This was replaced with a much simpler design,
described in the final version of the PMF paper, in which all passwords above a
length are considered valid on the basis that they were probably machine
generated.